+++ src/MSALWrapper.Test/MSALWrapper.Test.csproj
@@ -1,9 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup Condition="$([MSBuild]::IsOSPlatform('Windows'))">
-    <TargetFramework>net5.0-windows10.0.17763.0</TargetFramework>
-    <SupportedOSPlatformVersion>7</SupportedOSPlatformVersion>
-    <DefineConstants>PlatformWindows</DefineConstants>
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
 
     <IsPackable>false</IsPackable>
 
@@ -11,15 +9,7 @@
     <TreatWarningsAsErrors>True</TreatWarningsAsErrors>
     <GenerateDocumentationFile>True</GenerateDocumentationFile>
   </PropertyGroup>
-
-  <PropertyGroup Condition="$([MSBuild]::IsOSPlatform('OSX'))">
-    <TargetFramework>net6.0</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="$([MSBuild]::IsOSPlatform('Linux'))">
-    <TargetFramework>net6.0</TargetFramework>
-  </PropertyGroup>
-
+  
   <ItemGroup>
     <!-- Stylecop required items -->
     <AdditionalFiles Include="..\..\stylecop\stylecop.json" Link="stylecop.json" />
