+++ src/MSALWrapper/MSALWrapper.csproj
@@ -46,10 +46,4 @@
   <ItemGroup Condition=" '$(TargetFramework)' == 'net472' ">
     <PackageReference Include="Microsoft.Identity.Client.Desktop" Version="4.42.0" />
   </ItemGroup>
-  <ItemGroup>
-    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
-        <_Parameter1>$(AssemblyName).Test</_Parameter1>
-        <!-- We use the value of AssemblyName to declare the value of the attribute -->
-    </AssemblyAttribute>
-  </ItemGroup>
 </Project>
