+++ src/MSALWrapper/PCAWrapper.cs
@@ -5,8 +5,10 @@ namespace Microsoft.Authentication.MSALWrapper
 {
     using System;
     using System.Collections.Generic;
+    using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
+    using Microsoft.Extensions.Logging;
     using Microsoft.Identity.Client;
     using Microsoft.IdentityModel.JsonWebTokens;
 
@@ -83,6 +85,20 @@ namespace Microsoft.Authentication.MSALWrapper
         /// The <see cref="Task"/>.
         /// </returns>
         Task<TokenResult> GetTokenDeviceCodeAsync(IEnumerable<string> scopes, Func<DeviceCodeResult, Task> callback, CancellationToken cancellationToken);
+
+        /// <summary>
+        /// Tries to get an account using the preferred domain if provided.
+        /// </summary>
+        /// <param name="preferredDomain">The preferred domain.</param>
+        /// <returns>The <see cref="Task"/>.</returns>
+        Task<IAccount> TryGetAccountAsync(string preferredDomain = null);
+
+        /// <summary>
+        /// Tries to get a list of accounts using the preferred domain if provided.
+        /// </summary>
+        /// <param name="preferredDomain">The preferred domain.</param>
+        /// <returns>The <see cref="Task"/>.</returns>
+        Task<IList<IAccount>> TryToGetAccountsAsync(string preferredDomain = null);
     }
 
     /// <summary>
@@ -93,16 +109,26 @@ namespace Microsoft.Authentication.MSALWrapper
     /// </summary>
     public class PCAWrapper : IPCAWrapper
     {
+        private ILogger logger;
         private IPublicClientApplication pca;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="PCAWrapper"/> class.
         /// </summary>
-        /// <param name="pca">
-        /// The pca.
-        /// </param>
+        /// <param name="pca">The public client application instance.</param>
         public PCAWrapper(IPublicClientApplication pca)
         {
+            this.pca = pca;
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="PCAWrapper"/> class.
+        /// </summary>
+        /// <param name="logger">The logger.</param>
+        /// <param name="pca">The public client application instance.</param>
+        public PCAWrapper(ILogger logger, IPublicClientApplication pca)
+        {
+            this.logger = logger;
             this.pca = pca;
         }
 
@@ -196,6 +222,46 @@ namespace Microsoft.Authentication.MSALWrapper
         {
             AuthenticationResult result = await this.pca.AcquireTokenWithDeviceCode(scopes, callback).ExecuteAsync(cancellationToken).ConfigureAwait(false);
             return this.TokenResultOrThrow(result);
+        }
+
+        /// <summary>
+        /// Tries to get an account using the preferred domain if provided.
+        /// </summary>
+        /// <param name="preferredDomain">The preferred domain.</param>
+        /// <returns>The <see cref="Task"/>.</returns>
+        public async Task<IAccount> TryGetAccountAsync(string preferredDomain = null)
+        {
+            var accounts = await this.TryToGetAccountsAsync(preferredDomain);
+            var account = (accounts == null || accounts.Count() > 1) ? null : accounts.FirstOrDefault();
+            return account;
+        }
+
+        /// <summary>
+        /// Tries to get a list of accounts using the preferred domain if provided.
+        /// </summary>
+        /// <param name="preferredDomain">The preferred domain.</param>
+        /// <returns>The <see cref="Task"/>.</returns>
+        public async Task<IList<IAccount>> TryToGetAccountsAsync(string preferredDomain = null)
+        {
+            IEnumerable<IAccount> accounts = await this.pca.GetAccountsAsync().ConfigureAwait(false);
+            if (accounts == null)
+            {
+                return null;
+            }
+
+            this.logger.LogDebug($"Accounts found in cache: ({accounts.Count()}):");
+            this.logger.LogDebug(string.Join("\n", accounts.Select(a => a.Username)));
+
+            if (!string.IsNullOrWhiteSpace(preferredDomain))
+            {
+                this.logger.LogDebug($"Filtering cached accounts with preferred domain '{preferredDomain}'");
+                accounts = accounts.Where(eachAccount => eachAccount.Username.EndsWith(preferredDomain, StringComparison.OrdinalIgnoreCase));
+
+                this.logger.LogDebug($"Accounts found in cache after filtering: ({accounts.Count()}):");
+                this.logger.LogDebug(string.Join("\n", accounts.Select(a => a.Username)));
+            }
+
+            return accounts.ToList();
         }
 
         private TokenResult TokenResultOrThrow(AuthenticationResult result)
